refactor(client): replace codex (exec) backend with codex-app-server impl (#177)
Drop the per-task `codex exec` fork-exec backend in favor of a single
persistent `codex app-server` process driven over stdio JSON-RPC, keeping
the user-facing name `codex` so operators flip nothing in config. Also
delete the parallel `codex-app-server` backend introduced in #172 — only
one codex transport needs to ship.

What moves:

- `packages/client/src/backends/codex-app-server.ts` → `codex.ts`
  (replacing the old per-task impl). `mapPartsToCodexInput` and
  `parseCodexTurnUsageForOpenAICompat` (the openai-compat usage forward)
  are inlined so the new file is self-contained.
- `codex-app-server-rpc.ts` → `codex-rpc.ts`. Log tag `[codex-app-server]`
  → `[codex-rpc]`.
- `codex-app-server.test.ts` → `codex.test.ts`. Symbol renames only;
  test scenarios unchanged.
- `cli.ts`: drop the `case 'codex-app-server':` dispatch; the existing
  `case 'codex':` now wires to the new impl. Drop `extra_args` plumbing.
- `config.ts`: drop `'codex-app-server'` from `BackendsConfig` and the
  known-backends set; drop `extra_args` from `CodexBackendConfig`; add
  `approval_decision` to `CodexBackendConfig`.
- `server/card-resolver.ts` + `cards/codex-app-server.json`: deleted.
- Server + client `cards/codex.json`: description updated to reflect
  the persistent app-server transport; client card adds
  `application/json` input mode for parity with the (deleted) app-server
  card.
- `claude.ts` comment: `historyToInjectItems in codex-app-server.ts` →
  `historyToInjectItems in codex.ts`.

Behavior changes worth flagging:

- Per-task fork-exec isolation is gone. Documented in the changeset.
- `tool_call_history` now flows through `thread/inject_items` as native
  Responses API items, not as a `<tool_call_history>` JSON blob in the
  user prompt — eliminates the multi-turn re-call loop fixed in #176.
- Built-in `shell_tool` / `unified_exec` are disabled per-thread via
  `config.features` instead of `--disable …` argv (fix from #175).
- Concurrent same-`contextId` tasks are serialised through a per-context
  lock (app-server rejects a second `turn/start` while another is active
  on the same thread).
- openai-compat usage forwarding (#165) is preserved by inlining
  `parseCodexTurnUsageForOpenAICompat`. The new code defensively reads
  `turn.usage` (and `params.usage` as a fallback) on `turn/completed`;
  if codex app-server emits usage on a different notification we'll
  add a separate handler. Spec contract is best-effort: when usage is
  absent the gateway falls back to its local cl100k_base estimate.

Both `.changeset/codex-app-server-backend.md` and the new behavior are
unreleased (last published is 0.14.0), so this is a clean swap of an
unshipped feature — no operator migration needed.

Closes #177.
