+++ packages/client/src/config.test.ts
@@ -273,79 +273,6 @@ test('normalizeConfig trims string fields and drops whitespace-only', (t) => {
   });
 });
 
-test('normalizeConfig keeps backends.codex.extra_args when it is a string[] (#147)', (t) => {
-  const dir = mkdtempSync(join(tmpdir(), 'vicoop-cfg-codex-extra-'));
-  t.after(() => rmSync(dir, { recursive: true, force: true }));
-  const path = join(dir, 'config.json');
-  writeFileSync(
-    path,
-    JSON.stringify({
-      backends: { codex: { extra_args: ['--skip-git-repo-check'] } },
-    }),
-  );
-  assert.deepEqual(readConfig(path), {
-    backends: { codex: { extra_args: ['--skip-git-repo-check'] } },
-  });
-});
-
-test('normalizeConfig drops backends.codex.extra_args when it is not a string[]', (t) => {
-  const dir = mkdtempSync(join(tmpdir(), 'vicoop-cfg-codex-extra-bad-'));
-  t.after(() => rmSync(dir, { recursive: true, force: true }));
-  const path = join(dir, 'config.json');
-  writeFileSync(
-    path,
-    JSON.stringify({
-      backends: {
-        codex: {
-          cwd: '/work',
-          extra_args: '--skip-git-repo-check', // wrong type — dropped, the rest kept
-        },
-      },
-    }),
-  );
-  assert.deepEqual(readConfig(path), {
-    backends: { codex: { cwd: '/work' } },
-  });
-});
-
-test('normalizeConfig drops backends.codex.extra_args containing non-strings', (t) => {
-  const dir = mkdtempSync(join(tmpdir(), 'vicoop-cfg-codex-extra-mixed-'));
-  t.after(() => rmSync(dir, { recursive: true, force: true }));
-  const path = join(dir, 'config.json');
-  writeFileSync(
-    path,
-    JSON.stringify({
-      backends: { codex: { extra_args: ['--ok', 123] } },
-    }),
-  );
-  assert.deepEqual(readConfig(path), {});
-});
-
-test('normalizeConfig trims each codex.extra_args entry and drops whitespace-only ones', (t) => {
-  // Matches the same trimming/whitespace-only rule used for other string
-  // config fields (server_url, server_token, …). A hand-edited
-  // `" --flag"` would otherwise reach codex argv with a leading space
-  // and produce a confusing parse error.
-  const dir = mkdtempSync(join(tmpdir(), 'vicoop-cfg-codex-extra-trim-'));
-  t.after(() => rmSync(dir, { recursive: true, force: true }));
-  const path = join(dir, 'config.json');
-  writeFileSync(
-    path,
-    JSON.stringify({
-      backends: {
-        codex: {
-          extra_args: ['  --skip-git-repo-check  ', '   ', '--debug'],
-        },
-      },
-    }),
-  );
-  assert.deepEqual(readConfig(path), {
-    backends: {
-      codex: { extra_args: ['--skip-git-repo-check', '--debug'] },
-    },
-  });
-});
-
 test('normalizeConfig drops unknown backend and unknown codex.sandbox_mode', (t) => {
   // Permissive normalize must also catch enum typos — otherwise a hand-edited
   // `"backend": "claud"` or `"sandbox_mode": "workspace_write"` would survive
@@ -374,17 +301,17 @@ test('normalizeConfig drops unknown backend and unknown codex.sandbox_mode', (t)
   });
 });
 
-test('normalizeConfig accepts backends.codex-app-server with sandbox_mode and approval_decision', (t) => {
-  const dir = mkdtempSync(join(tmpdir(), 'vicoop-cfg-cas-'));
+test('normalizeConfig accepts backends.codex with sandbox_mode and approval_decision', (t) => {
+  const dir = mkdtempSync(join(tmpdir(), 'vicoop-cfg-codex-approval-'));
   t.after(() => rmSync(dir, { recursive: true, force: true }));
   const path = join(dir, 'config.json');
   writeFileSync(
     path,
     JSON.stringify({
       server_url: 'wss://b',
-      backend: 'codex-app-server',
+      backend: 'codex',
       backends: {
-        'codex-app-server': {
+        codex: {
           cwd: '/srv',
           sandbox_mode: 'workspace-write',
           approval_decision: 'acceptForSession',
@@ -394,9 +321,9 @@ test('normalizeConfig accepts backends.codex-app-server with sandbox_mode and ap
   );
   assert.deepEqual(readConfig(path), {
     server_url: 'wss://b',
-    backend: 'codex-app-server',
+    backend: 'codex',
     backends: {
-      'codex-app-server': {
+      codex: {
         cwd: '/srv',
         sandbox_mode: 'workspace-write',
         approval_decision: 'acceptForSession',
@@ -405,17 +332,16 @@ test('normalizeConfig accepts backends.codex-app-server with sandbox_mode and ap
   });
 });
 
-test('normalizeConfig drops invalid codex-app-server approval_decision and sandbox_mode', (t) => {
-  const dir = mkdtempSync(join(tmpdir(), 'vicoop-cfg-cas-bad-'));
+test('normalizeConfig drops invalid codex approval_decision', (t) => {
+  const dir = mkdtempSync(join(tmpdir(), 'vicoop-cfg-codex-approval-bad-'));
   t.after(() => rmSync(dir, { recursive: true, force: true }));
   const path = join(dir, 'config.json');
   writeFileSync(
     path,
     JSON.stringify({
       backends: {
-        'codex-app-server': {
+        codex: {
           cwd: '/work',
-          sandbox_mode: 'workspace_write', // typo — dropped
           approval_decision: 'yes', // not in enum — dropped
         },
       },
@@ -423,7 +349,7 @@ test('normalizeConfig drops invalid codex-app-server approval_decision and sandb
   );
   assert.deepEqual(readConfig(path), {
     backends: {
-      'codex-app-server': { cwd: '/work' },
+      codex: { cwd: '/work' },
     },
   });
 });
