--- packages/server/src/cards/codex-app-server.json
@@ -1,40 +1,0 @@
-{
-  "name": "codex-app-server",
-  "description": "OpenAI Codex agent driven via `codex app-server` (persistent stdio JSON-RPC). One long-lived codex process serves all turns; drop-in capability surface for the `codex` backend. Accepts text, inline image, and JSON data part inputs.",
-  "version": "0.0.1",
-  "protocolVersion": "0.3.0",
-  "capabilities": {
-    "streaming": true,
-    "extensions": [
-      {
-        "uri": "https://github.com/a2aproject/a2a-samples/extensions/traceability/v1",
-        "description": "Opt-in execution trace stream for Codex command executions.",
-        "required": false
-      },
-      {
-        "uri": "https://github.com/planetarium/oai2a2a/extensions/openai-compat/v1",
-        "description": "Honors OpenAI-style system/tools/tool_choice via message metadata; emulates tool calls by instructing the model to emit a {\"tool_calls\":[...]} JSON envelope, returned as an A2A data part.",
-        "required": false
-      }
-    ]
-  },
-  "defaultInputModes": [
-    "text/plain",
-    "image/png",
-    "image/jpeg",
-    "image/gif",
-    "image/webp",
-    "application/json"
-  ],
-  "defaultOutputModes": [
-    "text/plain"
-  ],
-  "skills": [
-    {
-      "id": "chat",
-      "name": "chat",
-      "description": "Ask a question or give an instruction in natural language; the Codex agent responds with text. Include image (image/{png,jpeg,gif,webp}) parts as inline `file.bytes`. Auxiliary structured context can be attached as A2A `data` parts; they are serialized into the prompt as tagged JSON blocks.",
-      "tags": ["chat", "assistant", "codex"]
-    }
-  ]
-}
